remove native modals

+++ app/javascript/components/Photos.js
@@ -3,10 +3,12 @@ import Gallery from "react-photo-gallery";
 import Carousel, { Modal, ModalGateway } from "react-images";
 import Header from './Header';
 import Image from './Image';
+import DeleteDialog from "./DeleteDialog";
 
 function Photos({photos, onDelete, onLike}) {
   const [currentImage, setCurrentImage] = useState(0);
   const [viewerIsOpen, setViewerIsOpen] = useState(false);
+  const [dialogIsOpen, setDialogIsOpen] = useState(false);
 
   const openLightbox = useCallback(index => {
     setCurrentImage(index);
@@ -18,20 +20,27 @@ function Photos({photos, onDelete, onLike}) {
     setViewerIsOpen(false);
   }, []);
 
-  const deleteCurrentPhoto = useCallback(() => {
-    onDelete(currentImage);
-    closeLightbox();
-  }, [currentImage, closeLightbox, onDelete]);
+  const deleteCurrentPhoto = useCallback((confirmed) => {
+    setDialogIsOpen(false);
+    if (confirmed) {
+      onDelete(currentImage);
+    }
+  }, [currentImage, onDelete]);
 
   const likeCurrentPhoto = useCallback(() => {
     onLike(photos[currentImage]);
     closeLightbox();
   }, [currentImage, closeLightbox, onLike, photos]);
 
+  const handleDialogOpen = () => {
+    closeLightbox();
+    setDialogIsOpen(true);
+  }
+
   const viewer = viewerIsOpen ? (
     <Modal onClose={closeLightbox}>
       <Carousel
-        components={{ Footer: null, NavigationPrev: null, Navigation: null, Header: Header(deleteCurrentPhoto, likeCurrentPhoto) }}
+        components={{ Footer: null, NavigationPrev: null, Navigation: null, Header: Header(handleDialogOpen, likeCurrentPhoto) }}
         currentIndex={currentImage}
         views={photos.map((x, idx) => ({
           ...x,
@@ -45,6 +54,7 @@ function Photos({photos, onDelete, onLike}) {
 
   return (
     <div>
+      <DeleteDialog image={photos[currentImage]} isOpen={dialogIsOpen} onClick={deleteCurrentPhoto}/>
       <Gallery photos={photos} onClick={openLightbox} renderImage={Image(openLightbox)}/>
       <ModalGateway>
         {viewer}
